- Commit after Tomasz branch reintegration (08/02/2012)

+++ api/cpp/cppapi/branches/NextRelease/server/w_attribute.cpp
@@ -4283,7 +4283,7 @@ void WAttribute::set_max_value(Tango::DevUChar &new_val)
 //
 
 	TangoSys_MemStream str;
-	str << new_val;
+	str << (short)new_val; // to store the numeric value
 	max_value_str = str.str();
 
 //
@@ -4841,7 +4841,7 @@ void WAttribute::set_min_value(Tango::DevUChar &new_val)
 //
 
 	TangoSys_MemStream str;
-	str << new_val;
+	str << (short)new_val; // to store the numeric value
 	min_value_str = str.str();
 
 //
